Move components to pure folder
