+++ src/cow-react/common/pure/ApproveButton/index.cosmos.tsx
